pronto

+++ main.py
@@ -6,8 +6,8 @@ metros_quadrados = float(metros_quadrados)
 
 # As duas variáveis qtd_de_latas e valor_total devem guardar a resposta do problema
 # Troque os zeros pelos valores corretos.
-qtd_de_latas = metros_quadrados // 54
-qtd_de_latas2 = metros_quadrados / 54
+qtd_de_latas = int(metros_quadrados // 54)
+qtd_de_latas2 = int(metros_quadrados / 54)
 valor_total = 0
 if qtd_de_latas2 > qtd_de_latas :
     lt = int(qtd_de_latas + 1)
